Module http_signatures::prelude

This module defines useful traits for using HTTP Signatures.

Traits

AsHttpSignature
AsHttpSignature defines a trait for getting an Authorization or Signature Header string from any type that implements it. It provides three methods: as_http_signature, which implementors must define, and authorization_header and signature_header, which use as_http_signature to create the header string.

GetKey
The GetKey trait is used during HTTP Signature verification to access the required decryption key based on a given key_id.

VerifyHeader
The VerifyHeader trait is meant to be implemented for the request types from http libraries (such as Hyper and Rocket). This trait makes verifying requests much easier, since the verify_authorization_header() and verify_signature_header() methods can be called directly on a Request type.

WithHttpSignature
WithHttpSignature defines a trait for adding Authorization and Signature headers to another library's request or response object.