Trait http_signatures::prelude::AsHttpSignature

pub trait AsHttpSignature<T> where
    T: Read, {
    fn as_http_signature(
        &self, 
        key_id: String, 
        key: T, 
        algorithm: SignatureAlgorithm
    ) -> Result<HttpSignature<T>, Error>;

    fn authorization_header(
        &self, 
        key_id: String, 
        key: T, 
        algorithm: SignatureAlgorithm
    ) -> Result<String, Error> { ... }

    fn signature_header(
        &self, 
        key_id: String, 
        key: T, 
        algorithm: SignatureAlgorithm
    ) -> Result<String, Error> { ... }
}

AsHttpSignature defines a trait for getting an Authorization or Signature Header string from any type that implements it. It provides three methods: as_http_signature, which implementors must define, and authorization_header and signature_header, which use as_http_signature to create the header string.

Required Methods

fn as_http_signature(
&self,
key_id: String,
key: T,
algorithm: SignatureAlgorithm
) -> Result<HttpSignature<T>, Error>

Gets an HttpSignature struct from an immutably borrowed Self

Provided Methods

fn authorization_header(
&self,
key_id: String,
key: T,
algorithm: SignatureAlgorithm
) -> Result<String, Error>

Generates the Authorization Header from an immutably borrowed Self

fn signature_header(
&self,
key_id: String,
key: T,
algorithm: SignatureAlgorithm
) -> Result<String, Error>

Generates the Signature Header from an immutably borrowed Self