Trait http_signatures::prelude::VerifyHeader

pub trait VerifyHeader {
    fn verify_signature_header<G: GetKey>(
        &self, 
        key_getter: G
    ) -> Result<(), VerificationError>;

    fn verify_authorization_header<G: GetKey>(
        &self, 
        key_getter: G
    ) -> Result<(), VerificationError>;
}

The VerifyHeader trait is meant to be implemented for the request types from http libraries (such as Hyper and Rocket). This trait makes verifying requests much easier, since the verify_authorization_header() and verify_signature_header() methods can be called directly on a Request type.

For examples, see the hyper server and rocket files.

Required Methods

fn verify_signature_header<G: GetKey>(
&self,
key_getter: G
) -> Result<(), VerificationError>

fn verify_authorization_header<G: GetKey>(
&self,
key_getter: G
) -> Result<(), VerificationError>