Struct http_signatures::SignedHeader

pub struct SignedHeader<'a> { /* fields omitted */ }

The SignedHeader struct is the direct reasult of reading in the Authorization or Signature header from a given request.

It contains the keys to the request's headers in the correct order for recreating the signing string, the algorithm used to create the signature, and the signature itself.

It also contains the key_id, which will be handled by a type implementing GetKey.

Methods

impl<'a> SignedHeader<'a>

pub fn new(s: &'a str) -> Result<Self, DecodeError>

Try to create an SignedHeader from a given String.

pub fn verify<G>(
self,
headers: &[(&str, &str)],
method: &str,
path: &str,
query: Option<&str>,
key_getter: G
) -> Result<(), VerificationError> where
G: GetKey,

Try to verify the current SignedHeader.