Module http_signatures::use_reqwest

Available with the use_reqwest feature. This module defines AsHttpSignature and WithHttpSignature for reqwest::Request.

Example adding a signature to a request

This example adds the HTTP Signature to the request directly as an Authorization header. with_signature_header can be used to add the signature as a Signature header instead.

let key = File::open("tests/assets/private.der")?;
let alg = SignatureAlgorithm::RSA(ShaSize::TwoFiftySix);

let client = Client::new();
let mut req = client.get("https://example.com").build()?;

req.with_authorization_header("rsa-key-1".into(), key, alg)?;

See this example for more usage information.